move built-in plugin dust collector to its own crate.
