+++ otx-plugin-protocol/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 ckb-types = "0.105"
+crossbeam-channel = "0.5.1"
 log = "0.4.17"
 serde = { version = "1.0", features = ["derive"] }
 serde_derive = "1.0"
