+++ otx-pool/Cargo.toml
@@ -31,3 +31,5 @@ tokio = { version = "1", features = ["sync", "process"] }
 otx-format = { path = "../otx-format" }
 otx-plugin-protocol = { path = "../otx-plugin-protocol" }
 utils = { path = "../utils" }
+
+dust-collector = { path = "../plugins-built-in/dust-collector" }
