+++ otx-pool/src/built_in_plugin/mod.rs
@@ -1,7 +1,5 @@
 mod atomic_swap;
-pub mod dust_collector;
 mod signer;
 
 pub use atomic_swap::AtomicSwap;
-pub use dust_collector::DustCollector;
 pub use signer::Signer;
