+++ plugins-built-in/dust-collector/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "dust-collector"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0"
+ckb-sdk = "=2.3.0"
+ckb-sdk-open-tx = { git = "https://github.com/EthanYuan/ckb-sdk-rust.git", branch = "opentx_sign_tx_ethan", package = "ckb-sdk"}
+ckb-stop-handler = "0.105"
+ckb-types = "0.105"
+dashmap = "5.4.0"
+derive_more = "0.99"
+log = "0.4.17"
+
+otx-format = { path = "../../otx-format" }
+otx-plugin-protocol = { path = "../../otx-plugin-protocol" }
+utils = { path = "../../utils" }
